Refactored cli_test
