+++ cli_test/helpers/constants.go
@@ -1,0 +1,13 @@
+package helpers
+
+const (
+	Denom        = "stake"
+	KeyFoo       = "foo"
+	KeyBar       = "bar"
+	FooDenom     = "footoken"
+	FeeDenom     = "feetoken"
+	Fee2Denom    = "fee2token"
+	KeyBaz       = "baz"
+	KeyVesting   = "vesting"
+	KeyFooBarBaz = "foobarbaz"
+)
