+++ cli_test/helpers/utils.go
@@ -1,0 +1,10 @@
+package helpers
+
+import "strings"
+
+func addFlags(cmd string, flags []string) string {
+	for _, f := range flags {
+		cmd += " " + f
+	}
+	return strings.TrimSpace(cmd)
+}
