--- cli_test/test_helpers.go
@@ -1,360 +1,0 @@
-package clitest
-
-import (
-	"encoding/json"
-	"fmt"
-	clientkeys "github.com/cosmos/cosmos-sdk/client/keys"
-	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/crypto/keyring"
-	"github.com/cosmos/cosmos-sdk/server"
-	"github.com/cosmos/cosmos-sdk/simapp"
-	"github.com/cosmos/cosmos-sdk/tests"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/stretchr/testify/require"
-	tmtypes "github.com/tendermint/tendermint/types"
-	"io/ioutil"
-	"os"
-	"path/filepath"
-	"strings"
-	"testing"
-	"time"
-)
-
-const (
-	denom        = "stake"
-	keyFoo       = "foo"
-	keyBar       = "bar"
-	fooDenom     = "footoken"
-	feeDenom     = "feetoken"
-	fee2Denom    = "fee2token"
-	keyBaz       = "baz"
-	keyVesting   = "vesting"
-	keyFooBarBaz = "foobarbaz"
-)
-
-var (
-	totalCoins = sdk.NewCoins(
-		sdk.NewCoin(fee2Denom, sdk.TokensFromConsensusPower(2000000)),
-		sdk.NewCoin(feeDenom, sdk.TokensFromConsensusPower(2000000)),
-		sdk.NewCoin(fooDenom, sdk.TokensFromConsensusPower(2000)),
-		sdk.NewCoin(denom, sdk.TokensFromConsensusPower(300).Add(sdk.NewInt(12))), // add coins from inflation
-	)
-
-	startCoins = sdk.NewCoins(
-		sdk.NewCoin(fee2Denom, sdk.TokensFromConsensusPower(1000000)),
-		sdk.NewCoin(feeDenom, sdk.TokensFromConsensusPower(1000000)),
-		sdk.NewCoin(fooDenom, sdk.TokensFromConsensusPower(1000)),
-		sdk.NewCoin(denom, sdk.TokensFromConsensusPower(150)),
-	)
-
-	vestingCoins = sdk.NewCoins(
-		sdk.NewCoin(feeDenom, sdk.TokensFromConsensusPower(500000)),
-	)
-)
-
-//___________________________________________________________________________________
-// Fixtures
-
-// Fixtures is used to setup the testing environment
-type Fixtures struct {
-	BuildDir     string
-	RootDir      string
-	SimdBinary   string
-	SimcliBinary string
-	ChainID      string
-	RPCAddr      string
-	Port         string
-	SimdHome     string
-	SimcliHome   string
-	P2PAddr      string
-	T            *testing.T
-}
-
-// NewFixtures creates a new instance of Fixtures with many vars set
-func NewFixtures(t *testing.T) *Fixtures {
-	tmpDir, err := ioutil.TempDir("", "sdk_integration_"+t.Name()+"_")
-	require.NoError(t, err)
-
-	servAddr, port, err := server.FreeTCPAddr()
-	require.NoError(t, err)
-
-	p2pAddr, _, err := server.FreeTCPAddr()
-	require.NoError(t, err)
-
-	buildDir := os.Getenv("BUILDDIR")
-	if buildDir == "" {
-		buildDir, err = filepath.Abs("../build/")
-		require.NoError(t, err)
-	}
-
-	return &Fixtures{
-		T:            t,
-		BuildDir:     buildDir,
-		RootDir:      tmpDir,
-		SimdBinary:   filepath.Join(buildDir, "simd"),
-		SimcliBinary: filepath.Join(buildDir, "simcli"),
-		SimdHome:     filepath.Join(tmpDir, ".simd"),
-		SimcliHome:   filepath.Join(tmpDir, ".simcli"),
-		RPCAddr:      servAddr,
-		P2PAddr:      p2pAddr,
-		Port:         port,
-	}
-}
-
-// GenesisFile returns the path of the genesis file
-func (f Fixtures) GenesisFile() string {
-	return filepath.Join(f.SimdHome, "config", "genesis.json")
-}
-
-// GenesisFile returns the application's genesis state
-func (f Fixtures) GenesisState() simapp.GenesisState {
-	cdc := codec.New()
-	genDoc, err := tmtypes.GenesisDocFromFile(f.GenesisFile())
-	require.NoError(f.T, err)
-
-	var appState simapp.GenesisState
-	require.NoError(f.T, cdc.UnmarshalJSON(genDoc.AppState, &appState))
-	return appState
-}
-
-// InitFixtures is called at the beginning of a test  and initializes a chain
-// with 1 validator.
-func InitFixtures(t *testing.T) (f *Fixtures) {
-	f = NewFixtures(t)
-
-	// reset test state
-	f.UnsafeResetAll()
-
-	f.CLIConfig("keyring-backend", "test")
-
-	// ensure keystore has foo and bar keys
-	f.KeysDelete(keyFoo)
-	f.KeysDelete(keyBar)
-	f.KeysDelete(keyBar)
-	f.KeysDelete(keyFooBarBaz)
-	f.KeysAdd(keyFoo)
-	f.KeysAdd(keyBar)
-	f.KeysAdd(keyBaz)
-	f.KeysAdd(keyVesting)
-	f.KeysAdd(keyFooBarBaz, "--multisig-threshold=2", fmt.Sprintf(
-		"--multisig=%s,%s,%s", keyFoo, keyBar, keyBaz))
-
-	// ensure that CLI output is in JSON format
-	f.CLIConfig("output", "json")
-
-	// NOTE: GDInit sets the ChainID
-	f.GDInit(keyFoo)
-
-	f.CLIConfig("chain-id", f.ChainID)
-	f.CLIConfig("broadcast-mode", "block")
-	f.CLIConfig("trust-node", "true")
-
-	// start an account with tokens
-	f.AddGenesisAccount(f.KeyAddress(keyFoo), startCoins)
-	f.AddGenesisAccount(
-		f.KeyAddress(keyVesting), startCoins,
-		fmt.Sprintf("--vesting-amount=%s", vestingCoins),
-		fmt.Sprintf("--vesting-start-time=%d", time.Now().UTC().UnixNano()),
-		fmt.Sprintf("--vesting-end-time=%d", time.Now().Add(60*time.Second).UTC().UnixNano()),
-	)
-
-	f.GenTx(keyFoo)
-	f.CollectGenTxs()
-
-	return f
-}
-
-// Cleanup is meant to be run at the end of a test to clean up an remaining test state
-func (f *Fixtures) Cleanup(dirs ...string) {
-	clean := append(dirs, f.RootDir)
-	for _, d := range clean {
-		require.NoError(f.T, os.RemoveAll(d))
-	}
-}
-
-// Flags returns the flags necessary for making most CLI calls
-func (f *Fixtures) Flags() string {
-	return fmt.Sprintf("--home=%s --node=%s", f.SimcliHome, f.RPCAddr)
-}
-
-//___________________________________________________________________________________
-// gaiad
-
-// UnsafeResetAll is gaiad unsafe-reset-all
-func (f *Fixtures) UnsafeResetAll(flags ...string) {
-	cmd := fmt.Sprintf("%s --home=%s unsafe-reset-all", f.SimdBinary, f.SimdHome)
-	executeWrite(f.T, addFlags(cmd, flags))
-	err := os.RemoveAll(filepath.Join(f.SimdHome, "config", "gentx"))
-	require.NoError(f.T, err)
-}
-
-// GDInit is gaiad init
-// NOTE: GDInit sets the ChainID for the Fixtures instance
-func (f *Fixtures) GDInit(moniker string, flags ...string) {
-	cmd := fmt.Sprintf("%s init -o --home=%s %s", f.SimdBinary, f.SimdHome, moniker)
-	_, stderr := tests.ExecuteT(f.T, addFlags(cmd, flags), clientkeys.DefaultKeyPass)
-
-	var chainID string
-	var initRes map[string]json.RawMessage
-
-	err := json.Unmarshal([]byte(stderr), &initRes)
-	require.NoError(f.T, err)
-
-	err = json.Unmarshal(initRes["chain_id"], &chainID)
-	require.NoError(f.T, err)
-
-	f.ChainID = chainID
-}
-
-// AddGenesisAccount is gaiad add-genesis-account
-func (f *Fixtures) AddGenesisAccount(address sdk.AccAddress, coins sdk.Coins, flags ...string) {
-	cmd := fmt.Sprintf("%s add-genesis-account %s %s --home=%s --keyring-backend=test", f.SimdBinary, address, coins, f.SimdHome)
-	executeWriteCheckErr(f.T, addFlags(cmd, flags))
-}
-
-// GenTx is gaiad gentx
-func (f *Fixtures) GenTx(name string, flags ...string) {
-	cmd := fmt.Sprintf("%s gentx --name=%s --home=%s --home-client=%s --keyring-backend=test", f.SimdBinary, name, f.SimdHome, f.SimcliHome)
-	executeWriteCheckErr(f.T, addFlags(cmd, flags))
-}
-
-// CollectGenTxs is gaiad collect-gentxs
-func (f *Fixtures) CollectGenTxs(flags ...string) {
-	cmd := fmt.Sprintf("%s collect-gentxs --home=%s", f.SimdBinary, f.SimdHome)
-	executeWriteCheckErr(f.T, addFlags(cmd, flags))
-}
-
-// GDStart runs gaiad start with the appropriate flags and returns a process
-func (f *Fixtures) GDStart(flags ...string) *tests.Process {
-	cmd := fmt.Sprintf("%s start --home=%s --rpc.laddr=%v --p2p.laddr=%v", f.SimdBinary, f.SimdHome, f.RPCAddr, f.P2PAddr)
-	proc := tests.GoExecuteTWithStdout(f.T, addFlags(cmd, flags))
-	tests.WaitForTMStart(f.Port)
-	tests.WaitForNextNBlocksTM(1, f.Port)
-	return proc
-}
-
-// GDTendermint returns the results of gaiad tendermint [query]
-func (f *Fixtures) GDTendermint(query string) string {
-	cmd := fmt.Sprintf("%s tendermint %s --home=%s", f.SimdBinary, query, f.SimdHome)
-	success, stdout, stderr := executeWriteRetStdStreams(f.T, cmd)
-	require.Empty(f.T, stderr)
-	require.True(f.T, success)
-	return strings.TrimSpace(stdout)
-}
-
-// ValidateGenesis runs gaiad validate-genesis
-func (f *Fixtures) ValidateGenesis() {
-	cmd := fmt.Sprintf("%s validate-genesis --home=%s", f.SimdBinary, f.SimdHome)
-	executeWriteCheckErr(f.T, cmd)
-}
-
-//___________________________________________________________________________________
-// gaiacli keys
-
-// KeysDelete is gaiacli keys delete
-func (f *Fixtures) KeysDelete(name string, flags ...string) {
-	cmd := fmt.Sprintf("%s keys delete --keyring-backend=test --home=%s %s", f.SimcliBinary,
-		f.SimcliHome, name)
-	executeWrite(f.T, addFlags(cmd, append(append(flags, "-y"), "-f")))
-}
-
-// KeysAdd is gaiacli keys add
-func (f *Fixtures) KeysAdd(name string, flags ...string) {
-	cmd := fmt.Sprintf("%s keys add --keyring-backend=test --home=%s %s", f.SimcliBinary,
-		f.SimcliHome, name)
-	executeWriteCheckErr(f.T, addFlags(cmd, flags))
-}
-
-// KeysAddRecover prepares gaiacli keys add --recover
-func (f *Fixtures) KeysAddRecover(name, mnemonic string, flags ...string) (exitSuccess bool, stdout, stderr string) {
-	cmd := fmt.Sprintf("%s keys add --keyring-backend=test --home=%s --recover %s",
-		f.SimcliBinary, f.SimcliHome, name)
-	return executeWriteRetStdStreams(f.T, addFlags(cmd, flags), mnemonic)
-}
-
-// KeysAddRecoverHDPath prepares gaiacli keys add --recover --account --index
-func (f *Fixtures) KeysAddRecoverHDPath(name, mnemonic string, account uint32, index uint32, flags ...string) {
-	cmd := fmt.Sprintf("%s keys add --keyring-backend=test --home=%s --recover %s --account %d"+
-		" --index %d", f.SimcliBinary, f.SimcliHome, name, account, index)
-	executeWriteCheckErr(f.T, addFlags(cmd, flags), mnemonic)
-}
-
-// KeysShow is gaiacli keys show
-func (f *Fixtures) KeysShow(name string, flags ...string) keyring.KeyOutput {
-	cmd := fmt.Sprintf("%s keys show --keyring-backend=test --home=%s %s", f.SimcliBinary,
-		f.SimcliHome, name)
-	out, _ := tests.ExecuteT(f.T, addFlags(cmd, flags), "")
-	var ko keyring.KeyOutput
-	err := clientkeys.UnmarshalJSON([]byte(out), &ko)
-	require.NoError(f.T, err)
-	return ko
-}
-
-// KeyAddress returns the SDK account address from the key
-func (f *Fixtures) KeyAddress(name string) sdk.AccAddress {
-	ko := f.KeysShow(name)
-	accAddr, err := sdk.AccAddressFromBech32(ko.Address)
-	require.NoError(f.T, err)
-	return accAddr
-}
-
-//___________________________________________________________________________________
-// gaiacli config
-
-// CLIConfig is gaiacli config
-func (f *Fixtures) CLIConfig(key, value string, flags ...string) {
-	cmd := fmt.Sprintf("%s config --home=%s %s %s", f.SimcliBinary, f.SimcliHome, key, value)
-	executeWriteCheckErr(f.T, addFlags(cmd, flags))
-}
-
-//___________________________________________________________________________________
-// executors
-
-func executeWriteCheckErr(t *testing.T, cmdStr string, writes ...string) {
-	require.True(t, executeWrite(t, cmdStr, writes...))
-}
-
-func executeWrite(t *testing.T, cmdStr string, writes ...string) (exitSuccess bool) {
-	exitSuccess, _, _ = executeWriteRetStdStreams(t, cmdStr, writes...)
-	return
-}
-
-func executeWriteRetStdStreams(t *testing.T, cmdStr string, writes ...string) (bool, string, string) {
-	proc := tests.GoExecuteT(t, cmdStr)
-
-	// Enables use of interactive commands
-	for _, write := range writes {
-		_, err := proc.StdinPipe.Write([]byte(write + "\n"))
-		require.NoError(t, err)
-	}
-
-	// Read both stdout and stderr from the process
-	stdout, stderr, err := proc.ReadAll()
-	if err != nil {
-		fmt.Println("Err on proc.ReadAll()", err, cmdStr)
-	}
-
-	// Log output.
-	if len(stdout) > 0 {
-		t.Log("Stdout:", string(stdout))
-	}
-	if len(stderr) > 0 {
-		t.Log("Stderr:", string(stderr))
-	}
-
-	// Wait for process to exit
-	proc.Wait()
-
-	// Return succes, stdout, stderr
-	return proc.ExitState.Success(), string(stdout), string(stderr)
-}
-
-//___________________________________________________________________________________
-// utils
-
-func addFlags(cmd string, flags []string) string {
-	for _, f := range flags {
-		cmd += " " + f
-	}
-	return strings.TrimSpace(cmd)
-}
